server/sonar-db-migration/src/main/java/org/sonar/server/platform/db/migration/version/v82/DeleteQgateConditionsUsingSecurityHotspotMetrics.java
/*
 * SonarQube
 * Copyright (C) 2009-2020 SonarSource SA
 * mailto:info AT sonarsource DOT com
 *
 * This program is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 3 of the License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public License
 * along with this program; if not, write to the Free Software Foundation,
 * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
package org.sonar.server.platform.db.migration.version.v82;

import java.sql.SQLException;
import org.sonar.db.Database;
import org.sonar.server.platform.db.migration.step.DataChange;
import org.sonar.server.platform.db.migration.step.MassUpdate;

public class DeleteQgateConditionsUsingSecurityHotspotMetrics extends DataChange {

  public DeleteQgateConditionsUsingSecurityHotspotMetrics(Database db) {
    super(db);
  }

  @Override
  protected void execute(Context context) throws SQLException {
    MassUpdate massUpdate = context.prepareMassUpdate();
    massUpdate.select("select qgc.id from quality_gate_conditions qgc " +
      "  inner join metrics m on m.id = qgc.metric_id and " +
      "    m.name in ('security_hotspots', 'new_security_hotspots')");
    massUpdate.update("delete from quality_gate_conditions where id = ?");
    massUpdate.execute((row, handler) -> {
      handler.setLong(1, row.getLong(1));
      return true;
    });
  }

}
